Use the DTO `RemoteAPODItem` so that there are less arrow references pointing to `GalleryImage` struct! Move the mapping RemoteAPODItem -> GalleryImage to `RemoteGalleryLoader`!
